改进立即数的edge case算法，现在不再依赖ub 修复JmpLike指令在操作数接近INT_MIN, INT_MAX时存在一个错误 添加更多寄存器 添加syscall, jle, jne指令 添加所有JmpLike的ret版本（如re表示return if equals） 统一内存架构 改进后优化器

+++ src/cpp/ir/ops/Mov.h
@@ -28,7 +28,7 @@ public:
         if (const auto &immediate = INSTANCEOF_SHARED(right, Immediate)) {
             if (const auto &result = INSTANCEOF_SHARED(left, Register)) {
                 return fmt::format("scoreboard players set {} vm_regs {}",
-                                   result->getName(), immediate->getValue());
+                                   result->getName(), static_cast<i32>(immediate->getValue()));
             } else {
                 assert(INSTANCEOF_SHARED(left, Ptr));
